feature: initial

+++ MasjidOnline.Data.Model/Captcha/CaptchaQuestionForAnswer.cs
@@ -2,7 +2,7 @@
 
 public class CaptchaQuestionForAnswer
 {
-    public required long Id { get; set; }
+    public required int Id { get; set; }
 
     public required float Degree { get; set; }
 }
