+++ MasjidOnline.Entity/Captcha/CaptchaAnswer.cs
@@ -4,11 +4,11 @@ namespace MasjidOnline.Entity.Captcha;
 
 public class CaptchaAnswer
 {
-    public required long Id { get; set; }
+    public required int Id { get; set; }
 
     public required DateTime DateTime { get; set; }
 
-    public required long CaptchaQuestionId { get; set; }
+    public required int CaptchaQuestionId { get; set; }
 
     public required float Degree { get; set; }
