+++ MasjidOnline.Entity/Captcha/CaptchaQuestion.cs
@@ -4,7 +4,7 @@ namespace MasjidOnline.Entity.Captcha;
 
 public class CaptchaQuestion
 {
-    public required long Id { get; set; }
+    public required int Id { get; set; }
 
     public required DateTime DateTime { get; set; }
