+++ MasjidOnline.Entity/Log/Exception.cs
@@ -1,10 +1,11 @@
 ﻿using System;
 
+// todo rename to MasjidOnline.Entity.EventLog
 namespace MasjidOnline.Entity.Log;
 
 public class Exception
 {
-    public required long Id { get; set; }
+    public required int Id { get; set; }
 
     public required DateTime DateTime { get; set; }
