+++ MasjidOnline.Entity/Payments/Payment.cs
@@ -4,13 +4,13 @@ namespace MasjidOnline.Entity.Payments;
 
 public class Payment
 {
-    public required long Id { get; set; }
+    public required int Id { get; set; }
 
     public required DateTime DateTime { get; set; }
 
-    public required long TransactionId { get; set; }
+    public required int TransactionId { get; set; }
 
     public required PaymentStatus PaymentStatus { get; set; }
 
-    public long UserId { get; set; }
+    public int UserId { get; set; }
 }
