+++ MasjidOnline.Entity/Transactions/TransactionFile.cs
@@ -2,7 +2,7 @@
 
 public class TransactionFile
 {
-    public required long Id { get; set; }
+    public required int Id { get; set; }
 
-    public required long TransactionId { get; set; }
+    public required int TransactionId { get; set; }
 }
